refactor(tools): Repair the types the conversion warnings named
-Wconversion and -Wsign-conversion reported forty-one implicit
conversions. Each one is answered by the type rather than by a cast,
because a cast records the answer without stating the question.

Eleven of them were one expression, repeated: the byte offset of a
pixel, written as `((size_t)y * im->w + x) * 4`. The first factor
carried the conversion and the other two arrived implicitly.
`img_offset` in img.h now holds that arithmetic, converts all three
terms, and leaves the call site reading as what it is.

The channel count and the bytes for each pixel index a buffer, so both
take the type of an index at their declaration, which removes the casts
that stood at each use. The products of two dimensions convert both
factors instead of one. The candidate of the hash chain keeps int,
because the chain ends with -1, and converts once inside the loop where
that bound already holds.

Three conversions from long to double sit in the ground colour of a
stamp. The divisor now says double, which is the arithmetic that the
expression already performed, because the dividend is a double.

The generators write the same bytes into every asset.

+++ tools/check.c
@@ -85,7 +85,7 @@ static int cell_width(
         if (i + (size_t)bytes > len) {
             break;
         }
-        for (int k = 1; k < bytes; k++) {
+        for (size_t k = 1; k < (size_t)bytes; k++) {
             cp = (cp << 6) | ((unsigned char)s[i + k] & 0x3F);
         }
         i += (size_t)bytes;
